refactor lib/widgets/stat_item.dart location
