+++ lib/widgets/stat_item.dart
